+++ R/server-side.R
@@ -3,7 +3,8 @@
 #' @description
 #' Change the state of the howler player from the server.
 #'
-#' \code{playHowl}, \code{pauseHowl} and \code{stopHowl} will all be applied to the current track.
+#' \code{playHowl}, \code{pauseHowl}, \code{togglePlayHowl} and \code{stopHowl}
+#' will all be applied to the current track.
 #'
 #' \code{changeTrack} will update the track to the file specified.
 #'
@@ -51,13 +52,23 @@ changeTrack <- function(id, track, session = getDefaultReactiveDomain()) {
 #' @export
 addTrack <- function(id, track, play_track = FALSE, session = getDefaultReactiveDomain()) {
   if (is.null(names(track))) {
-    track_name <- sub("\\.[^\\.]+$", "", basename(track[1]))
+    track_name <- vapply(
+      track,
+      function(x) sub("\\.[^\\.]+$", "", basename(x[1])),
+      character(1),
+      USE.NAMES = FALSE
+    )
   } else {
     track_name <- names(track)
   }
 
   message_name <- paste0("addHowlerTrack_", session$ns(id))
-  session$sendCustomMessage(message_name, list(track = track, track_name = track_name, play = play_track))
+  track_info <- list(
+    track = as.list(unname(track)),
+    track_name = as.list(track_name),
+    play = play_track
+  )
+  session$sendCustomMessage(message_name, track_info)
 }
 
 #' @rdname howlerServer
@@ -71,6 +82,13 @@ playHowl <- function(id, session = getDefaultReactiveDomain()) {
 #' @export
 pauseHowl <- function(id, session = getDefaultReactiveDomain()) {
   message_name <- paste0("pauseHowler_", session$ns(id))
+  session$sendCustomMessage(message_name, id)
+}
+
+#' @rdname howlerServer
+#' @export
+togglePlayHowl <- function(id, session = getDefaultReactiveDomain()) {
+  message_name <- paste0("togglePlayHowler_", session$ns(id))
   session$sendCustomMessage(message_name, id)
 }